backend/requirements.txt
Django==3.1.7
psycopg2>=2.8
markdown
django-filter
djangorestframework
django-cors-headers
Pillow
djangorestframework-simplejwt
importlib_metadata==8.4.0
gunicorn
python-decouple
requests
django-redis
daphne
channels
channels-redis==4.1.0
